Reword the README.
I'm not sure if there are specific conventions for
Jenkins plugin readmes, but I'm taking my best shot here.

I found the README to be a bit hard to understand,
so after setting it up on my own Jenkins instance,
I've updated the README to reflect what exists both
in the config settings and for general readability.