Create missing PR during synchronize trigger
When the GitHub syncronize trigger is sent (e.g. adding commits to a
pull request) it can fail if Jenkins didn't previously know about the
PR.  This can happen, for example, if Jenkins missed an `opened`
message, or if the PR was created before GHPRB was enabled.

This fixes this by creating the pull request if it is missing.