Use the correct SHA1 for Matrix builds
The problem was that job.getBuilds() would return the ongoing matrix
root build, causing to build the previous SHA1 in case a commit was not
the first of the pull request.

The current fix breaks the “smart changes” logic for matrix jobs.
A proper fix that preserves this functionality because it would be very
hard to determine which is a root build of the same run, especially if
the concurrent builds options is enabled.

The PR also simplifies some logic.