Changes getProjectName to check for null from the trigger's getActualProject()
This isn't a bug per-se, but I have a Jenkins instance configured using the job-dsl plugin.  There was an attempt to configure the ghprb triggers using a configure block (which directly edits xml) which created a somewhat invalid trigger.  There was an error processing the webhooks in GhprbRootAction.doIndex().   The try catch block in the loop caught the exception and should have continued processing the rest of the hooks for other projects, however the catch block calls getProjectName, which has the potential to throw NullPointerException if the project for the trigger can't be found.  Avoid this by checking for whether trigger.getActualProject() return null.