Pull PR data if necessary when determining mergeability
I fixed a bug in the Java Github API where it would always populate certain fields of the structure regardless of whether they had been pulled before (the case is unmerged PR).  Doing so causes the mergeability check to fail because we will never pull new data.  Instead, if the mergeability is null, grab the PR again.

Alternatively, I think this functionality should be removed altogether.  It's expensive and in most cases the build will immediately start and we will make another API to get the mergeability data again.  I will file another PR for that.