Queue status extensions
This change adds a "queue status" extension to the plugin.  What this does is update the github status for a job based on the number of items in the jenkins queue that are assigned the same label.  This is useful for systems where the number of queue items can become large, so that users have some idea of when their job will run.  Of course, this doesn't take into account queue vetoing, etc. plugins.  When enabled, the github contexts are periodically updated with a " X/Y in <label>" suffix on the build status description.