Update list of current maintainers
The intent of this change is to reflect in Jenkins wiki metadata the
current active maintainers.  This is so Jenkins users have an up to date
list of points of contact for this plugin.

For now, I've removed @janinko and @valdisrigdon since I haven't seen
them post in a while.  If you guys don't want to be removed, then comment
in this PR and let me know you want your ID to remain.  I'll wait ~1 week
before merging this for your feedback.

@jenkinsci/ghprb-plugin-developers after this change is merged I
recommend you following up with PRs to add yourselves if you wish to be
reflected as a point of contact in the Jenkins wiki.