Cancel downstream builds if was triggered by GHPRB.
On the situation where a build trigger a new async builds, if any user
updates the PR only one build will be cancel, the upstream one.

When downstreams jobs take a lot to run, the CI is block with builds
taht does not make sense to have running at all.

With this change, each build is validated if the cause is a
`UpstreamCause` and if it is true, this validates that the parent build
was triggered by GHPRB, if so will be cancel if the PR id match.

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>