Deprecate GitHub Pull Request Builder plugin
The GitHub Pull Request Builder plugin should be deprecated so that
users do not mistakenly install it when integrating with GitHub.

More details are available in Jenkins infrastructure help desk ticket:

* https://github.com/jenkins-infra/helpdesk/issues/4608

The documentation for plugins.jenkins.io is being updated at:

* https://github.com/jenkins-infra/plugins-wiki-docs/pull/8

The deprecation message is being added here as one more location that
declares the plugin is deprecated.